Updated NavBar, Hero Image Size, new Out & About Page!
I have modified the nav bar to be responsive to hover when facilities is hovered over, make the nav links to facilities work.

There is now an Out & About page containing image, description and google maps locations of attractions. Images and a find out more link to direct users to the attractions