Booking and Payment Pages Started
I have created a booking and a payment page, there are some temporary in-line JS to disable form submission and lead to payment page. Additionally some minor changes to make the BookNow look better. There is a sample image added which needs to be replaced with the image of the hotel room when the rooms page and booking functionality is complete.